Refactored name resolution functions into mixin
