+++ src/codemodder/codemods/https_connection.py
@@ -1,11 +1,7 @@
 from libcst import matchers
 from libcst.codemod.visitors import AddImportsVisitor, RemoveImportsVisitor
-from libcst.helpers import get_full_name_for_node
 from libcst.metadata import (
-    Assignment,
-    ImportAssignment,
     PositionProvider,
-    ScopeProvider,
 )
 from codemodder.codemods.base_codemod import (
     BaseCodemod,
@@ -13,6 +9,7 @@ from codemodder.codemods.base_codemod import (
     ReviewGuidance,
 )
 from codemodder.codemods.change import Change
+from codemodder.codemods.utils_mixin import NameResolutionMixin
 from codemodder.file_context import FileContext
 import libcst as cst
 from libcst.codemod import (
@@ -30,13 +27,12 @@ class HTTPSConnection(BaseCodemod, Codemod):
     )
     CHANGE_DESCRIPTION = "Enforced HTTPS connection"
 
-    METADATA_DEPENDENCIES = (PositionProvider, ScopeProvider)
+    METADATA_DEPENDENCIES = (PositionProvider,)
 
     matching_functions = {
         "urllib3.HTTPConnectionPool",
         "urllib3.connectionpool.HTTPConnectionPool",
     }
-    replacing_function = "HTTPSConnectionPool"
 
     def __init__(self, codemod_context: CodemodContext, file_context: FileContext):
         Codemod.__init__(self, codemod_context)
@@ -50,8 +46,8 @@ class HTTPSConnection(BaseCodemod, Codemod):
         return visitor.transform_module(tree)
 
 
-class ConnectionPollVisitor(VisitorBasedCodemodCommand):
-    METADATA_DEPENDENCIES = (PositionProvider, ScopeProvider)
+class ConnectionPollVisitor(VisitorBasedCodemodCommand, NameResolutionMixin):
+    METADATA_DEPENDENCIES = (PositionProvider,)
 
     def __init__(self, codemod_context: CodemodContext, file_context: FileContext):
         super().__init__(codemod_context)
@@ -63,9 +59,9 @@ class ConnectionPollVisitor(VisitorBasedCodemodCommand):
         pos_to_match = self.node_position(original_node)
         line_number = pos_to_match.start.line
         if self.filter_by_path_includes_or_excludes(pos_to_match):
-            true_name = self._get_full_true_name(original_node.func)
+            true_name = self.find_base_name(original_node.func)
             if (
-                self.is_call_from_imported_module(original_node)
+                self._is_direct_call_from_imported_module(original_node)
                 and true_name in HTTPSConnection.matching_functions
             ):
                 self.changes_in_file.append(
@@ -91,80 +87,6 @@ class ConnectionPollVisitor(VisitorBasedCodemodCommand):
                 )
         return updated_node
 
-    def _get_full_true_name(self, node):
-        if matchers.matches(node, matchers.Name()):
-            maybe_assignment = self.find_single_assignment(node)
-            if maybe_assignment and isinstance(maybe_assignment, ImportAssignment):
-                import_node = maybe_assignment.node
-                for alias in import_node.names:
-                    if maybe_assignment.name in (
-                        alias.evaluated_alias,
-                        alias.evaluated_name,
-                    ):
-                        return self._get_true_name_for_import(import_node, alias)
-            return node.value
-
-        if matchers.matches(node, matchers.Attribute()):
-            maybe_name = self._get_full_true_name(node.value)
-            if maybe_name:
-                return maybe_name + "." + node.attr.value
-        return None
-
-    def _get_true_name_for_import(self, import_node, import_alias):
-        """
-        For import nodes, this is defined as the module name. For ImportFrom, this is defined as <module name>.<alias_name>.
-        """
-        if matchers.matches(import_node, matchers.Import()):
-            return get_full_name_for_node(import_alias.name)
-        # it is a from import
-        return _get_name(import_node) + "." + get_full_name_for_node(import_alias.name)
-
-    def is_call_from_imported_module(
-        self, call: cst.Call
-    ) -> tuple[cst.Import | cst.ImportFrom, cst.ImportAlias] | None:
-        for nodo in iterate_left_expressions(call):
-            if matchers.matches(nodo, matchers.Name() | matchers.Attribute()):
-                maybe_assignment = self.find_single_assignment(nodo)
-                if isinstance(maybe_assignment, ImportAssignment):
-                    import_node = maybe_assignment.node
-                    for alias in import_node.names:
-                        if maybe_assignment.name in (
-                            alias.evaluated_alias,
-                            alias.evaluated_name,
-                        ):
-                            return (import_node, alias)
-        return None
-
-    def _find_assignments(
-        self, node: cst.Name | cst.Attribute | cst.Call | cst.Subscript | cst.Decorator
-    ) -> set[Assignment]:
-        """
-        Given a MetadataWrapper and a CSTNode with a possible access to it, find all the possible assignments that it refers.
-        """
-        scope = self.get_metadata(ScopeProvider, node)
-        # TODO workaround for a bug in libcst
-        if matchers.matches(node, cst.Attribute):
-            for access in scope.accesses:
-                if access.node == node:
-                    # pylint: disable=protected-access
-                    return access._Access__assignments
-        else:
-            if node in scope.accesses:
-                # pylint: disable=protected-access
-                return next(iter(scope.accesses[node]))._Access__assignments
-        return set()
-
-    def find_single_assignment(
-        self, node: cst.Name | cst.Attribute
-    ) -> Assignment | None:
-        """
-        Given a MetadataWrapper and a CSTNode representing an access, find if there is a single assignment that it refers to.
-        """
-        assignments = self._find_assignments(node)
-        if len(assignments) == 1:
-            return next(iter(assignments))
-        return None
-
     def filter_by_path_includes_or_excludes(self, pos_to_match):
         """
         Returns False if the node, whose position in the file is pos_to_match, matches any of the lines specified in the path-includes or path-excludes flags.
@@ -181,36 +103,5 @@ class ConnectionPollVisitor(VisitorBasedCodemodCommand):
         return self.get_metadata(PositionProvider, node)
 
 
-def iterate_left_expressions(node: cst.BaseExpression):
-    yield node
-    if matchers.matches(node, matchers.Attribute()):
-        yield from iterate_left_expressions(node.value)
-    if matchers.matches(node, matchers.Call()):
-        yield from iterate_left_expressions(node.func)
-
-
-def get_leftmost_expression(node: cst.BaseExpression) -> cst.BaseExpression:
-    """
-    Given an expression with dots (e.g. a.b.call()), extract the leftmost expression.
-    """
-    if matchers.matches(node, matchers.Attribute()):
-        return get_leftmost_expression(node.value)
-    if matchers.matches(node, matchers.Call(func=matchers.Attribute())):
-        return get_leftmost_expression(node.func)
-    return node
-
-
 def match_line(pos, line):
     return pos.start.line == line and pos.end.line == line
-
-
-def _get_name(node):
-    """
-    Get the full name of a module referenced by a Import or ImportFrom node.
-    For relative modules, dots are added at the beginning
-    """
-    if matchers.matches(node, matchers.ImportFrom()):
-        return "." * len(node.relative) + (get_full_name_for_node(node.module) or "")
-    if matchers.matches(node, matchers.Import()):
-        return get_full_name_for_node(node.names[0].name)
-    return ""
